Move class util.Typedef into typemap
Along with functions typedef_wrapped_defaults and lookup_c_statements.
